Edlin con IO
Modifique el archivo edlin para que usara IO, además de organizar el código por carpetas, agregar imagenes y los modelosUML
